+++ SwiftI18n/Classes/Views/SwiftUI/LocText.swift
@@ -7,68 +7,6 @@
 
 import SwiftUI
 
-private struct AcessibilityLabelModifier: ViewModifier {
-    @State private var language = I18nManager.instance.language
-    let key: String
-
-    func body(content: Content) -> some View {
-        if #available(iOS 14.0, *) {
-            content
-                .accessibilityLabel(Text(I18nManager.instance.localizedString(forKey: key, language: language)))
-                .modifier(LanguageModifier(didUpdatedLanguage: { language = $0 }))
-        } else {
-            content
-                .accessibility(label: Text(I18nManager.instance.localizedString(forKey: key, language: language)))
-                .modifier(LanguageModifier(didUpdatedLanguage: { language = $0 }))
-        }
-    }
-}
-
-private struct AcessibilityHintModifier: ViewModifier {
-    @State private var language = I18nManager.instance.language
-    let key: String
-
-    func body(content: Content) -> some View {
-        if #available(iOS 14.0, *) {
-            content
-                .accessibilityHint(Text(I18nManager.instance.localizedString(forKey: key, language: language)))
-                .modifier(LanguageModifier(didUpdatedLanguage: { language = $0 }))
-        } else {
-            content
-                .accessibility(hint: Text(I18nManager.instance.localizedString(forKey: key, language: language)))
-                .modifier(LanguageModifier(didUpdatedLanguage: { language = $0 }))
-        }
-    }
-}
-
-public extension View {
-    func setAcessibilityLabel(key: String) -> some View {
-        modifier(AcessibilityLabelModifier(key: key))
-    }
-
-    func setAcessibilityHint(key: String) -> some View {
-        modifier(AcessibilityHintModifier(key: key))
-    }
-}
-
-public struct LanguageModifier: ViewModifier {
-
-    private let didUpdatedLanguage: (String) -> Void
-
-    public init(didUpdatedLanguage: @escaping (String) -> Void){
-        self.didUpdatedLanguage = didUpdatedLanguage
-    }
-
-    public func body(content: Content) -> some View {
-        content
-            .onReceive(
-                NotificationCenter.default.publisher(for: .loc_LanguageDidChangeNotification).receive(on: DispatchQueue.main)
-            ) { _ in
-                didUpdatedLanguage(I18nManager.instance.language)
-            }
-    }
-}
-
 public struct LocText: View {
 
     fileprivate let content: [Content]
@@ -82,7 +20,7 @@ public struct LocText: View {
 
     public var body: some View {
         text
-            .modifier(LanguageModifier(didUpdatedLanguage: { language = $0 }))
+            .modifier(LanguageModifier(didUpdateLanguage: { language = $0 }))
             .environment(\.locale, Locale(identifier: language))
     }
 }
